lazily create our work_dir and cargo.toml file rather than every time the extension loads

+++ src/gucs.rs
@@ -22,50 +22,52 @@ pub(crate) fn init() {
         &PLRUST_PG_CONFIG,
         GucContext::Sighup,
     );
+}
 
-    // create our work directory if it doesn't exist
-    let work_dir = work_dir();
+pub(crate) fn work_dir() -> PathBuf {
+    let work_dir = PathBuf::from_str(
+        &PLRUST_WORK_DIR
+            .get()
+            .expect("plrust.work_dir is not set in postgresql.conf"),
+    )
+    .expect("plrust.work_dir is not a valid path");
+
+    // create the work dir if it doesn't exist
     if !work_dir.exists() {
         std::fs::create_dir_all(&work_dir)
             .expect("failed to create directory specified by plrust.work_dir");
     }
 
-    // bootstrap a configuration for pgx
-    let mut pgx_dir = work_dir;
-    pgx_dir.push(".pgx");
-    if !pgx_dir.exists() {
-        std::fs::create_dir_all(&pgx_dir)
-            .expect("failed to create pgx directory in plrust.work_dir");
-    }
-
-    let mut config_toml = pgx_dir;
-    config_toml.push("config.toml");
-    std::fs::write(
-        config_toml,
-        &format!(
-            r#"[configs]
-pg{}="{}"
-"#,
-            pg_sys::get_pg_major_version_string(),
-            pg_config()
-        ),
-    )
-    .expect("failed to write config.toml file");
-}
-
-pub(crate) fn work_dir() -> PathBuf {
-    PathBuf::from_str(
-        &PLRUST_WORK_DIR
-            .get()
-            .expect("plrust.work_dir is not set in postgresql.conf"),
-    )
-    .expect("plrust.work_dir is not a valid path")
+    work_dir
 }
 
 pub(crate) fn pgx_dir() -> PathBuf {
-    let mut path = work_dir();
-    path.push(".pgx");
-    path
+    let mut pgx_dir = work_dir();
+    pgx_dir.push(".pgx");
+
+    if !pgx_dir.exists() {
+        std::fs::create_dir_all(&pgx_dir)
+            .expect("failed to create `.pgx` directory in plrust.work_dir");
+    }
+
+    let mut config_toml = pgx_dir.clone();
+    config_toml.push("config.toml");
+    if !config_toml.exists() {
+        std::fs::write(
+            config_toml,
+            &format!(
+                r#"
+[configs]
+pg{}="{}"
+"#,
+                pg_sys::get_pg_major_version_string(),
+                pg_config()
+            ),
+        )
+        .expect("failed to write config.toml in plrust.work_dir/.pgx");
+    }
+
+    pgx_dir
 }
 
 fn pg_config() -> String {
